tiaAudio: updateSamples takes json text as argument
the previous version read the json textarea directly, which would be
inflexible for many applications

initialise() renamed to initialiseWithExampleJSON() to emphasise
that it's not needed for the tiaAudio engine to work

channel field in json renamed to control. the original json files had
misnamed this unintentionally

improved logging. including a message indicating the time taken for the
emulation to complete